move: rename SelectViewModel.kt to GenerateViewModel.kt
