+++ app/src/main/java/com/ry05k2ulv/myapplication/ui/MagApp.kt
@@ -1,9 +1,5 @@
 package com.ry05k2ulv.myapplication.ui
 
-import android.annotation.SuppressLint
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.RowScope
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Settings
@@ -11,9 +7,7 @@ import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.MediumTopAppBar
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.SmallTopAppBar
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarColors
@@ -24,8 +18,6 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.navigation.compose.rememberNavController
 import com.ry05k2ulv.myapplication.navigation.MagNavHost
 import com.ry05k2ulv.myapplication.ui.generate.select.SELECT_NAVIGATION_ROUTE
-import com.ry05k2ulv.myapplication.ui.generate.select.SelectScreen
-import com.ry05k2ulv.myapplication.ui.generate.select.SelectViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
